source/requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.19
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Django==1.11.23
django-cors-headers==3.0.2
django-redis==4.4.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
kombu==3.0.37
pycparser==2.19
PyJWT==1.7.1
pyspeedtest==1.2.7
pytz==2019.3
redis==2.10.5
requests==2.22.0
urllib3==1.25.6
